scripts/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
rich
python-dotenv
azure-identity
azure-storage-blob==12.23.1
azure-search-documents==11.6.0b7